requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
pydantic==2.9.2
pydantic-settings==2.4.0
python-dotenv==1.0.1
httpx==0.27.2
SQLAlchemy==2.0.35
aiosqlite==0.20.0
passlib[bcrypt]==1.7.4
itsdangerous==2.2.0
python-multipart==0.0.9
cryptography==43.0.1
